[X86][SSE] Fixed typo in insertps lowering.
We were checking for 2 insertions (which is caught earlier in the pattern matching loop) instead of the case where we have no insertions.

Turns out this code never fires as we always try to lower to insertps after trying to lower to blendps, which would catch these cases - I'm about to make some changes to support combining to insertps which could cause this to fire so I don't want to remove it.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@274648 91177308-0d34-0410-b5e6-96231b3b80d8