[X86] Make the FMA3 instruction names consistent between VEX and EVEX encoded versions.
This places the 132/213/231 form number in front of the SS/SD/PS/PD. Move the Y for 256-bit versions to be after the PS/PD. Change the AVX512 scalar forms to include a Z in the their name. This new format should be consistent with the general naming of instructions.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@276559 91177308-0d34-0410-b5e6-96231b3b80d8